requirements.txt
Flask==3.0.3
tensorflow-cpu==2.17.0
Pillow==10.4.0
numpy==1.26.4
streamlit==1.37.1
